# Conflicts:
#	Cargo.lock
#	packages/rs-platform-wallet/examples/basic_usage.rs
#	packages/rs-platform-wallet/src/platform_wallet_info/matured_transactions.rs
#	packages/rs-platform-wallet/src/platform_wallet_info/wallet_info_interface.rs
#	packages/rs-platform-wallet/src/platform_wallet_info/wallet_transaction_checker.rs

+++ packages/rs-dapi/src/clients/core_client.rs
@@ -90,7 +90,7 @@ impl CoreClient {
             .guarded_blocking_call(|client| client.get_block_count())
             .await??;
 
-        Ok(height as u32)
+        Ok(height)
     }
 
     /// Fetch verbose transaction metadata by txid hex string.
